\name{BigWigViews}
\alias{BigWigViews}
\title{BigWigViews class}
\arguments{
  \item{bigWigPaths}{character vector}

  \item{bigWigSamples}{DataFrame}

  \item{bigWigRanges}{GRanges}

  \item{bigWigExperiment}{a list}
}
\description{
  A lightweight pointer to a series of BigWigs and a set of
  GRanges over which to view the data.

  Constructor for BigWigViews objects
}

